requirements.txt
customtkinter
psutil
pystray
Pillow
pyyaml
Jinja2
